Moved files and improved folders structure
