+++ apollo/tools/core.py
@@ -11,7 +11,7 @@ import uuid
 from typing import Any
 import ollama
 
-from apollo.config.tools import get_available_tools
+from apollo.config.instructions import get_available_tools
 from apollo.config.const import Constant
 from apollo.service.format import format_duration_ns
 from apollo.service.session import save_user_history_to_json
